Clearing user posts relations

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const fs = require("fs");
 const path = require("path");
 const { validationResult } = require("express-validator");
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const page = req.query.page || 1;
@@ -18,7 +19,7 @@ exports.getPosts = (req, res, next) => {
     .then(posts => {
       res.status(200).json({ posts: posts, totalItems: totalItems });
     })
-    .catch(e => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -48,21 +49,31 @@ exports.createPost = (req, res, next) => {
     title: title,
     content: content,
     imageUrl: "images/vishwa.jpeg",
-    creator: {
-      name: "vish"
-    }
+    creator: req.userId
   });
+  let creator;
   post
     .save()
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: "Post created successfully",
-        post: result
+        post: post,
+        creator: { name: creator.name, id: creator._id }
       });
     })
     .catch(e => {
-      console.log(e);
+      if (!e.statusCode) {
+        e.statusCode = 500;
+      }
+      next(e);
     });
 };
 exports.getPost = (req, res, next) => {
@@ -97,6 +108,16 @@ exports.updatePost = (req, res, next) => {
 
   Post.findById(postId)
     .then(post => {
+      if (!post) {
+        const error = new Error("Post not found");
+        error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized");
+        error.statusCode = 403;
+        throw error;
+      }
       if (req.file) {
         imageUrl = req.file.path;
         if (imageUrl !== post.imageUrl) {
@@ -109,11 +130,6 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
-      if (!post) {
-        const error = new Error("Post not found");
-        error.statusCode = 404;
-        throw error;
-      }
       post.title = title;
       post.content = content;
       post.imageUrl = imageUrl;
@@ -124,7 +140,7 @@ exports.updatePost = (req, res, next) => {
         .status(200)
         .json({ message: "Post Updated successfully", post: result });
     })
-    .catch(e => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -140,14 +156,26 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       // look for loggen in user and author of the post
       return Post.findByIdAndDelete(postId);
     })
     .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Post deleted successfully" });
     })
-    .catch(e => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
